Refactor into separate header and source files
